+++ server/repository/src/db_diesel/mod.rs
@@ -93,8 +93,8 @@ mod patient;
 pub mod period;
 pub mod plugin_data;
 mod plugin_data_row;
-pub mod printer_configuration;
-pub mod printer_configuration_row;
+pub mod printer;
+pub mod printer_row;
 pub mod program_enrolment;
 mod program_enrolment_row;
 pub mod program_event;
@@ -223,7 +223,7 @@ pub use patient::*;
 pub use period::*;
 pub use plugin_data::*;
 pub use plugin_data_row::*;
-pub use printer_configuration_row::*;
+pub use printer_row::*;
 pub use program_enrolment::*;
 pub use program_enrolment_row::*;
 pub use program_event::*;
